Fix behavior of "auto" parser to be first parser returning links wins
This is how it was actually documented to work in the code, but had been
intentionally broken to just try to work with as much as possible. With more
rigorous testing of the parsers and add's behavior, we shouldn't need that.

History explained here:
https://github.com/ArchiveBox/ArchiveBox/issues/1363#issuecomment-1970294015

With this, we pass 9 of the 19 tests in tests/parser.